+++ CHANGELOG.md
@@ -2,8 +2,10 @@
 
 ## [Unreleased][unreleased]
 
-- New API methods for Database:
-  - async count(table, condition): number
+## [1.4.0][] - 2021-08-09
+
+- New API methods for Database `async count(table, condition): number`
+- Added support for functions in select (min, avg, count...)
 
 ## [1.3.0][] - 2021-08-07
 
@@ -87,7 +89,8 @@
 
 Code before fork from https://github.com/metarhia/sql
 
-[unreleased]: https://github.com/metarhia/metasql/compare/v1.3.0...HEAD
+[unreleased]: https://github.com/metarhia/metasql/compare/v1.4.0...HEAD
+[1.4.0]: https://github.com/metarhia/metasql/compare/v1.3.0...v1.4.0
 [1.3.0]: https://github.com/metarhia/metasql/compare/v1.2.0...v1.3.0
 [1.2.0]: https://github.com/metarhia/metasql/compare/v1.1.5...v1.2.0
 [1.1.5]: https://github.com/metarhia/metasql/compare/v1.1.4...v1.1.5